requirements.txt
fastmcp>=2.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
PyMySQL==1.1.0
openai>=1.3.0
instructor>=1.0.0
pydantic>=2.5.0
asyncio-pool==0.6.0
python-dotenv>=1.0.0
sqlparse>=0.4.0
greenlet>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
python-Levenshtein>=0.21.0